Move util/trace_replay.* to trace_replay/
Summary: util/ means for lower level libraries. trace_replay is highly integrated to DB and sometimes call DB. Move it out to a separate directory.

Test Plan: Run both of make and cmake
