+++ TARGETS
@@ -221,6 +221,7 @@ cpp_library(
         "tools/ldb_cmd.cc",
         "tools/ldb_tool.cc",
         "tools/sst_dump_tool.cc",
+        "trace_replay/trace_replay.cc",
         "util/bloom.cc",
         "util/build_version.cc",
         "util/coding.cc",
@@ -241,7 +242,6 @@ cpp_library(
         "util/string_util.cc",
         "util/thread_local.cc",
         "util/threadpool_imp.cc",
-        "util/trace_replay.cc",
         "util/xxhash.cc",
         "utilities/backupable/backupable_db.cc",
         "utilities/blob_db/blob_compaction_filter.cc",
