+++ trace_replay/trace_replay.cc
@@ -3,7 +3,7 @@
 //  COPYING file in the root directory) and Apache 2.0 License
 //  (found in the LICENSE.Apache file in the root directory).
 
-#include "util/trace_replay.h"
+#include "trace_replay/trace_replay.h"
 
 #include <chrono>
 #include <sstream>
